pyproject.toml
[project]
name = "foundry"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "numpy",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "ruff>=0.13.0"]
tools = ["matplotlib>=3.5", "pyyaml>=6.0"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "RET",  # flake8-return
]
ignore = [
    "SIM108",  # if-else-block-instead-of-if-exp
    "SIM117",  # combine-multiple-isinstance
    "E501",    # line-too-long (data often legitimately long)
    "N806",    # variable-name-lowercase (B, T, C, N, L, H, Q common in ML)
    "N803",    # argument-name-lowercase (X, Y common in ML)
    "N812",    # lowercase-imported-as-non-lowercase (F for functional)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
pythonpath = ["src"]
